TIDIJ-222 Refactored Firebase manipulation code into separate classes.
