+++ UDP/entities/utilities/strings.py
