+++ UDP/entities/vehicle/vehicle_statistics.py
